backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn==0.27.0

# Groq SDK
groq==0.11.0

# Tavily Search API
tavily-python==0.3.3

# Pydantic for structured outputs
pydantic==2.6.0

# Environment variables
python-dotenv==1.0.0

# CORS middleware
python-multipart==0.0.6
